refactor: ProductService 구조 리팩토링(#354)
